+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Removed support for payment links with `https://w3id.org/valueflows/Proposal` rel type.
 - Removed support for `MitraJcsRsaSignature2022` proof type.
 - Removed deprecated `reject-media` and `accept-media` filter actions.
+- Removed `read-outbox` command.
 
 ## [4.20.0] - 2026-03-19